% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_part_events}
\alias{seed_demo_part_events}
\title{Seed demo: journal des événements TSN/TSO (part_events)}
\usage{
seed_demo_part_events(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: journal des événements TSN/TSO (part_events)
}
