% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_parts_instances}
\alias{seed_demo_parts_instances}
\title{Seed demo: unités physiques individuelles (parts)}
\usage{
seed_demo_parts_instances(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: unités physiques individuelles (parts)
}
